Add a function to upload images in the groupController file
Import the Multer middleware to handle uploading images in the new groups field.
Create the path to the public folder to store the uploaded images of the users.
Generate an id name for images with the Shortid extension, to add unique id names plus the extension type.
Restrict non-image file types and manage errors.

Tags: Multer, Shortid, Nanoid