Finish CRUD operations in groups
Add the templates to edit groups with their respective routes and add
their methods in the groupController.
Authenticate the user before performing any editing operations.
Add the form and view to edit groups and save changes to the database.
Add the form and view to edit the new group image, load the new image
and remove the old one, use the fs unlink asynchronous method with the
image path to perform the operation.
Add the form and view to delete a group and delete the group and its
image also from the database and from the upload group folder
with the same fs unLink method.