Add maps
Add webpack and its configuration in the root folder. Add the script to the package.json.
Import the Leaflet scripts and styles and place the map with the map id in the new party view. Add the app.js file with the map settings. Add the leaflet-geosearch library to search and find addresses and locations. Set markers settings.