Finish CRUD operations for the party
Add moment.JS to display the party date and time in the admin view, and
group parties based on date. Add the DATEONLY data type to Sequelize
party model to drop the timestamp from the data. List the previous
parties under the group of groups. Fill in the form fields to edit the
pre-party information, including map and point coordinates.
Add the same map and Geoapify functionality in this field to edit
the address location. Add the form to delete a party with the same
template as the deleted group and add the functionality to delete the
party permanently i.e. in the admin views and in the database.
Install postgis to add a geometry data type to the location field in
the model, add the postgis CREATE EXTENSION query to the database.
Find problems trying to sanitize model inputs.
Changing the order of the model date that follows the ISO model does
not result in the type of data sequence in date that I want to change.