Add forms to edit profile and change password
Add the functionality to edit the profile and save it on the database.
Create a new section for changing the password.
Create a new prototype with the hash function of bcrypt to hash the
new password after passing the current password validation.
Update the user model hook to call the hashPassword prototype to avoid
hashing the password each time the user edits their profile.
After saving the new password on the database, call Passport Logout()
function and pass it the callback to redirect the user to the login area
after flashing a success message or returning an error with Next().