Add profile picture upload and user logout
Add the same groupController methods to load and update
the profile picture in the userController.
Add in the locals a method to see if the user is logged in or not.
Use Passport's logout() method to log the user out and redirect
them to the login page.
Add the partials with the authenticated user and unauthenticated user
paths to the layout.