Add styling to home view
Add the method to the homeController to display the next three parties
in the home view. Add the party model attributes, and include the
users and groups attributes to display their information on the home
view cards.
Add styles to the home view to correctly display the information of the
user and their created parties.