Add comments section
Add the comments model, controller, and methods for creating and deleting
comments.
Show alert messages with sweetalert2.

Validate the user before deleting the comments.
Loop through all the party comments and use axios after validating the
result to set or remove the action from the element by the authenticated
user of the session or the user who created the comment, and call the
controller function to destroy the comment from the database.

Fire scenario-based alert messages with response status with sweetalert2
to delete or not delete comments.