Add nearby parties
Add the nearbyParties method to the showParty function in the partyseekerController.
Check the location of the party.
Use the Sequelize.literal function to extract the unescaped geometry
data from the PostGIS ST_GeomFromText function with the party location
data points. Use the PostGIS function ST_DistanceSphere to generate the
distance in meters based on location data previously extracted from the
database.
Show a limit of three parties closest to the current party and
offset the first to not show the current party in the same view of
the current party that is displayed.
Sort the parties by distance.