Add party search
Add the seekerController with the method to search the results of the
home view inputs in the search area. Add the controller and its binding
to the routes.
Get the results of the data with req.query, use the GET method on the
search form action. Use the Sequelize operators to get the matching data
drawn from all parties.
Add the view for the search results, display the results, if any, in cards.