PartySeeker Deployment
Troubleshooting during deployment via Heroku.
Move dotenv from devDependencies to dependencies with
npm install dotenv with the --save-prod flag.
Lighten the size of the main images and minimize the CSS code. Limit the maximum upload size of images to 10 MB.
Add description meta.
Add the JavaScript files from the public directory to the Webpack entry point in an array.